B - upload preview modal could only open once
files can not be used in the anonymous function, because the following line clears vedleggElement, causing the files variable - a live object - to be cleared.

Since setState functions are called after the completion of the method this had the effect of only allowing the modal to be shown once.

By creating an array and passing it to the handler, we freeze the list and the problem is solved.